+++ src.ts/wordlists/wordlists-extra.ts
@@ -6,3 +6,5 @@ export { LangKo } from "./lang-ko.js";
 export { LangIt } from "./lang-it.js";
 export { LangPt } from "./lang-pt.js";
 export { LangZh } from "./lang-zh.js";
+export { LangSk } from "./lang-sk.js";
+export { LangDe } from "./lang-de.js";